outputView 구조체 생성
플레이어 이름을 출력하는 함수를 outputView구조체 생성하여 분리
