+++ LadderGame/LadderGame/LadderGame.swift
@@ -18,22 +18,4 @@ struct LadderGame {
             ResultView.showLadders(ladderFloorCount: ladderHeightCount, showLadders: ladders)
         }
     }
-    
-
-    //플레이어이름 출력하는 함수
-    func showPlayerName(playerName: Array<String>) {
-        var showName = String()
-        
-        for index in 0..<playerName.count {
-            var name = playerName[index]
-            while name.count < 6 {
-                name.insert(" ", at: name.endIndex)
-            }
-            showName = showName + name
-        }
-        print(showName)
-    }
-
-    
-    
 }
